chore: fix unit tests
