+++ src/features/common/services/cosmos-auth.ts
@@ -1,49 +1,34 @@
-import { CosmosClient } from "@azure/cosmos"
+import { CosmosClient, CosmosClientOptions } from "@azure/cosmos"
+import { DeviceCodeCredential, DefaultAzureCredential } from "@azure/identity"
 
 import logger from "@/features/insights/app-insights"
 
-const _cosmosAccessToken: string | null = null
-
 export const CONFIG = {
-  endpoint: process.env.APIM_BASE,
-  key: process.env.APIM_KEY,
+  endpoint: process.env.COSMOS_DB_ENDPOINT || "https://your-cosmos-account.documents.azure.com:443/",
   dbName: process.env.AZURE_COSMOSDB_DB_NAME || "localdev",
 }
 export type CosmosConfig = typeof CONFIG
 
-export const getCosmosAccessToken = async ({ endpoint, key }: CosmosConfig): Promise<string> => {
-  try {
-    const response = await fetch(`${endpoint}/cosmos`, {
-      method: "GET",
-      headers: { "api-key": key },
-      cache: "reload",
-    })
+let _cosmosClient: CosmosClient | null = null
 
-    if (!response.ok) {
-      logger.error(`🚀 > getCosmosAccessToken > ${response.status} ${response.statusText}`)
-      throw new Error(`${response.statusText}`)
+/**
+ * Create Cosmos Client with Azure AD Authentication
+ * @param config CosmosConfig
+ * @returns CosmosClient
+ */
+export function createCosmosClient(config: CosmosConfig = CONFIG): CosmosClient {
+  if (_cosmosClient) return _cosmosClient
+
+  try {
+    const options: CosmosClientOptions = {
+      endpoint: config.endpoint,
+      aadCredentials:
+        process.env.NODE_ENV === "development" ? new DeviceCodeCredential() : new DefaultAzureCredential(),
     }
-
-    const token = await response.text()
-    return token
+    _cosmosClient = new CosmosClient(options)
+    return _cosmosClient
   } catch (error) {
-    throw new Error(`Failed to fetch Cosmos Auth Token: ${JSON.stringify(error)}`)
+    logger.error(`Failed to create CosmosClient: ${JSON.stringify(error)}`)
+    throw new Error(`Failed to create CosmosClient: ${JSON.stringify(error)}`)
   }
-}
-
-export const isTokenExpired = (authToken: string | null = _cosmosAccessToken): boolean => {
-  try {
-    if (!authToken) return true
-    const expiry = JSON.parse(Buffer.from(authToken.split(".")[1], "base64").toString()).exp
-    const currentTime = Math.floor(Date.now() / 1000) // in seconds
-    return expiry <= currentTime
-  } catch (error) {
-    throw new Error(`Failed to check token expiry: ${error}`)
-  }
-}
-
-export async function createCosmosClient({ endpoint, key, ...rest }: CosmosConfig): Promise<CosmosClient> {
-  const authToken = await getCosmosAccessToken({ endpoint, key, ...rest })
-  const defaultHeaders = { "api-key": key, Authorization: `type=aad&ver=1.0&sig=${authToken}` }
-  return new CosmosClient({ endpoint, defaultHeaders })
 }
