+++ src/features/common/services/cosmos.ts
@@ -1,45 +1,6 @@
-import { CosmosClient, CosmosClientOptions, Container, PartitionKeyDefinition } from "@azure/cosmos"
-import { DefaultAzureCredential, DeviceCodeCredential } from "@azure/identity"
+import { Container, PartitionKeyDefinition } from "@azure/cosmos"
 
-import logger from "@/features/insights/app-insights"
-
-export const CONFIG = {
-  endpoint: process.env.COSMOS_DB_ENDPOINT || "https://your-cosmos-account.documents.azure.com:443/",
-  dbName: process.env.AZURE_COSMOSDB_DB_NAME || "localdev",
-}
-export type CosmosConfig = typeof CONFIG
-
-let cosmosClient: CosmosClient | null = null
-
-/**
- * Create Cosmos Client with Azure AD Authentication
- * @param config CosmosConfig
- * @returns CosmosClient
- */
-export function createCosmosClient(config: CosmosConfig): CosmosClient {
-  if (cosmosClient) return cosmosClient
-
-  try {
-    let credential
-
-    if (process.env.NODE_ENV === "development") {
-      credential = new DeviceCodeCredential()
-    } else {
-      credential = new DefaultAzureCredential()
-    }
-
-    const options: CosmosClientOptions = {
-      endpoint: config.endpoint,
-      aadCredentials: credential,
-    }
-
-    cosmosClient = new CosmosClient(options)
-    return cosmosClient
-  } catch (error) {
-    logger.error(`Failed to create CosmosClient: ${JSON.stringify(error)}`)
-    throw new Error(`Failed to create CosmosClient: ${JSON.stringify(error)}`)
-  }
-}
+import { CONFIG, CosmosConfig, createCosmosClient } from "./cosmos-auth"
 
 const _containerCache: Map<string, Container> = new Map()
